Revert "refactor: move mock signer to common package"
This reverts commit fe56e63a4758e4f6b95dedba65ea9a5c54ef0c7f.
