+++ consensus/integration/signer_test.go
@@ -1,1 +1,48 @@
 package integration_test
+
+import (
+	"github.com/onflow/flow-go/consensus/hotstuff/model"
+	"github.com/onflow/flow-go/model/flow"
+)
+
+type Signer struct {
+	localID flow.Identifier
+}
+
+func (*Signer) CreateProposal(block *model.Block) (*model.Proposal, error) {
+	proposal := &model.Proposal{
+		Block:   block,
+		SigData: nil,
+	}
+	return proposal, nil
+}
+func (s *Signer) CreateVote(block *model.Block) (*model.Vote, error) {
+	vote := &model.Vote{
+		View:     block.View,
+		BlockID:  block.BlockID,
+		SignerID: s.localID,
+		SigData:  nil,
+	}
+	return vote, nil
+}
+func (*Signer) CreateQC(votes []*model.Vote) (*flow.QuorumCertificate, error) {
+	voterIDs := make([]flow.Identifier, 0, len(votes))
+	for _, vote := range votes {
+		voterIDs = append(voterIDs, vote.SignerID)
+	}
+	qc := &flow.QuorumCertificate{
+		View:      votes[0].View,
+		BlockID:   votes[0].BlockID,
+		SignerIDs: voterIDs,
+		SigData:   nil,
+	}
+	return qc, nil
+}
+
+func (*Signer) VerifyVote(voterID *flow.Identity, sigData []byte, block *model.Block) (bool, error) {
+	return true, nil
+}
+
+func (*Signer) VerifyQC(voters flow.IdentityList, sigData []byte, block *model.Block) (bool, error) {
+	return true, nil
+}
